+++ index.php
@@ -1,9 +1,6 @@
 <?php
 
 use Kirby\Cms\App as Kirby;
-use Kirby\Filesystem\Dir;
-use Kirby\Filesystem\F;
-use Kirby\Toolkit\A;
 use Kirby\Toolkit\Str;
 use JanHerman\Utils\Translation;
 
@@ -21,14 +18,5 @@ Kirby::plugin('jan-herman/utils', [
     /**
      * Automatically load translations from `site/translations` folder
      */
-    'translations' => A::keyBy(
-        A::map(
-            Dir::read($dir = kirby()->root('languages') . '/translations'),
-            fn ($file) => A::merge(
-                ['lang' => F::name($file)],
-                Translation::load($dir . '/' . $file)
-            )
-        ),
-        'lang'
-    )
+    'translations' => Translation::loadDir(kirby()->root('languages') . '/translations'),
 ]);